remove the premature optimization of finding gpu objects before hashing make the `dfsCache` into a multimap to separate dfs from hashing and gpu object finding, get rid of `dfs_.._t`
Explanation:

To know the hash of an asset, one needs to know the its patch and the hash of its dependants ( and they need to know their patches, and so on).
Unfortunately while performing the DFS, the state of the user (i.e. buffer being used as SSBO in descriptor set) determines the patch of the dependant,
so we need to fully explore DFS to always have the patch of a dependant and re-hash properly.