ive been backed into a corner
[22.12.2025 19:53:22:605382][ERROR]: Descriptor set (000002B092D5C450) doesn't allow descriptor of such type category at binding 1.
[22.12.2025 19:53:22:605551][ERROR]: Invalid descriptor type (descriptorWrites[2]) [4c26069f9][nbl::video::ILogicalDevice::updateDescriptorSets - C:\Projects\Nabla\src\nbl\video\ILogicalDevice.cpp:529]

this is my error code at MeshRenderer.cpp Line 322

It coudl potentially be that the buffer just isn't flagged for storage buffer (it probably isnt), or it could be a deeper issue. My concern is that if I start tugging on this thread next thing I know I'l be rewriting all of nabla.
I'm just going to scrap the current geometry and write a simple cube and call it a day.

I verified that the tasks shaders worked correctly at some point during this process, whoevers reading this is just gonna have to take my word for it.